Moved localforage utilities to a new directory structure for better organization. Improved script loading with retry logic to enhance reliability in setting and injecting LocalForage during tests. Added the localforage.min.js script locally for consistent access.
Revised `run_pw.sh` to improve argument handling, including dynamic port usage and BASE_URL configuration. Simplified help messages and removed mandatory API_ENDPOINT checks. Updated GitHub workflow for compatibility with script changes.

(cherry picked from commit 920dbeb4148fc5133fb80144c81050b5a9f673ea)
