+++ e2etests/run_pw.sh
@@ -3,45 +3,28 @@
 # Default values
 CONFIG_FILE="playwright.regression.config.ts"
 UPDATE_SCREENSHOTS=false
-BASE_URL="http://0.0.0.0:4000"
 RUN_APP=false
 KEEP_RUNNING=false
 API_ENDPOINT=""
 CI_MODE=
+BASE_URL=""
+PORT=3000
+
 
 # Help message
 show_help() {
-    echo "Usage: $0 [API_ENDPOINT] [options]"
-    echo "Arguments:"
-    echo "  API_ENDPOINT            The API endpoint URL to use (optional if FFC_API_ENDPOINT env var is set)"
+    echo "Usage: $0 [options]"
     echo "Options:"
     echo "  -c, --config FILE        Playwright config file (default: regression-tests/playwright.regression.config.ts)"
     echo "  -u, --update             Update screenshots"
-    echo "  -U, --url URL           Base URL for the application (default: http://0.0.0.0:4000)"
-    echo "  -a, --run-application   Run the application in ngui folder"
+    echo "  -U, --url URL           Base URL for the application (default: $DEFAULT_BASE_URL)"
+    echo "  -p, --port PORT         Port to use for the application (default: $PORT)"
+    echo "  -a, --run-application API_ENDPOINT   Run the application in ngui folder with specified API endpoint"
     echo "  -k, --keep-running      Keep the application running after tests complete"
     echo "  -i, --ci                Set CI=true environment variable"
     echo "  -h, --help              Show this help message"
     exit 0
 }
-
-# Check if API_ENDPOINT is provided as first argument
-if [ $# -gt 0 ] && [[ ! "$1" =~ ^- ]]; then
-    API_ENDPOINT="$1"
-    shift
-fi
-
-# If API_ENDPOINT is not set via argument, try to get it from environment
-if [ -z "$API_ENDPOINT" ]; then
-    if [ -n "$FFC_API_ENDPOINT" ]; then
-        API_ENDPOINT="$FFC_API_ENDPOINT"
-    else
-        echo
-        echo "Error: API_ENDPOINT is required. Either provide it as an argument or set FFC_API_ENDPOINT environment variable."
-        echo
-        show_help
-    fi
-fi
 
 # Parse command line arguments
 while [[ $# -gt 0 ]]; do
@@ -58,9 +41,18 @@ while [[ $# -gt 0 ]]; do
             BASE_URL="$2"
             shift 2
             ;;
+        -p|--port)
+            PORT="$2"
+            shift 2
+            ;;
         -a|--run-application)
             RUN_APP=true
-            shift
+            if [ -z "$2" ] || [[ "$2" =~ ^- ]]; then
+                echo "Error: API_ENDPOINT is required when using -a/--run-application option"
+                show_help
+            fi
+            API_ENDPOINT="$2"
+            shift 2
             ;;
         -k|--keep-running)
             KEEP_RUNNING=true
@@ -80,34 +72,46 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
-# Check if -U and -a are used together
-if [ "$RUN_APP" = true ] && [ "$BASE_URL" != "http://0.0.0.0:4000" ]; then
-    echo "Error: Options -U/--url and -a/--run-application cannot be used together"
-    echo "When running the application (-a), the URL is fixed to http://0.0.0.0:4000"
-    exit 1
+# Set DEFAULT_BASE_URL with new port
+if [[ "$OSTYPE" == "darwin"* ]] || [[ "$OSTYPE" == "msys"* ]]; then
+    DEFAULT_BASE_URL="http://host.docker.internal:$PORT"
+else
+    DEFAULT_BASE_URL="http://0.0.0.0:$PORT"
+fi
+
+# Set BASE_URL based on options
+if [ "$RUN_APP" = true ]; then
+    if [ -n "$BASE_URL" ]; then
+        echo "Error: Options -U/--url and -a/--run-application cannot be used together"
+        echo "When running the application (-a), the URL is fixed to http://0.0.0.0:$PORT"
+        exit 1
+    fi
+    BASE_URL="http://0.0.0.0:$PORT"
+elif [ -z "$BASE_URL" ]; then
+    BASE_URL="$DEFAULT_BASE_URL"
 fi
 
 # Function to run tests in Docker container
 run_tests() {
     echo "Running tests in Docker container..."
-    
+
     # Build the test container
     docker build -t playwright-tests -f docker/Dockerfile.linux .
-    
+
     # Build test arguments
     TEST_ARGS="--config=$CONFIG_FILE"
-    
+
     # Add update screenshots flag if specified
     if [ "$UPDATE_SCREENSHOTS" = true ]; then
         TEST_ARGS="$TEST_ARGS --update-snapshots"
     fi
-    
+
     # Build environment variables
     ENV_ARGS="-e BASE_URL=$BASE_URL"
     if [ "$CI_MODE" = true ]; then
         ENV_ARGS="$ENV_ARGS -e CI=true"
     fi
-    
+
     # Run the tests
     docker run --rm \
         --network host \
@@ -130,21 +134,21 @@ cleanup() {
     elif [ "$KEEP_RUNNING" = true ]; then
         echo "Keeping application container running..."
         echo "Container name: ngui-container"
-        echo "Access URL: http://0.0.0.0:4000"
+        echo "Access URL: $BASE_URL"
     fi
 }
 
 # Function to run the application
 run_application() {
     echo "Starting ngui application..."
-    
+
     # Check if container exists and stop it
     if docker ps -a --format '{{.Names}}' | grep -q "^ngui-container$"; then
         echo "Stopping existing ngui-container..."
         docker stop ngui-container
         docker rm ngui-container
     fi
-    
+
     # Build environment variables
     ENV_ARGS="-e PROXY_URL=$API_ENDPOINT \
         -e KEEPER_ENDPOINT=$API_ENDPOINT \
@@ -153,17 +157,17 @@ run_application() {
         -e AUTH_ENDPOINT=$API_ENDPOINT \
         -e BUILD_MODE=production \
         -e UI_BUILD_PATH=/usr/src/app/ui"
-    
+
     if [ "$CI_MODE" = true ]; then
         ENV_ARGS="$ENV_ARGS -e CI=true"
     fi
-    
+
     # Start the application using Dockerfile
     docker build -t ngui-app -f ../ngui/Dockerfile ../.
-    docker run -d --name ngui-container -p 4000:4000 \
+    docker run -d --name ngui-container -p $PORT:4000 \
         $ENV_ARGS \
         ngui-app
-    
+
     # Wait for container to be running with timeout
     echo "Waiting for container to start..."
     TIMEOUT=30
@@ -178,13 +182,13 @@ run_application() {
             exit 1
         fi
     done
-    
+
     # Wait for the application to be ready
-    echo "Waiting for application to be ready..."
-    until curl -s http://0.0.0.0:4000 > /dev/null; do
+    echo "Waiting for application at $BASE_URL to be ready..."
+    until curl -s $BASE_URL > /dev/null; do
         sleep 1
     done
-    
+
     echo "Application is ready!"
 }
 
@@ -198,4 +202,3 @@ fi
 
 # Run tests
 run_tests
-
